chore: fixes linting issues
