+++ .github/workflows/deploy_website.yml
@@ -25,7 +25,7 @@ jobs:
           main-branch-name: 'develop'
       - uses: actions/setup-node@v3
         with:
-          node-version: 16.18.1
+          node-version: 18.13.0
           cache: yarn
       - name: Install dependencies
         run: yarn install --immutable
@@ -78,7 +78,7 @@ jobs:
           main-branch-name: 'develop'
       - uses: actions/setup-node@v3
         with:
-          node-version: 16.18.1
+          node-version: 18.13.0
           cache: yarn
       - name: Install dependencies
         run: yarn install --immutable
