+++ .github/workflows/publish.yaml
@@ -10,7 +10,7 @@ jobs:
       - uses: actions/checkout@v3
       - uses: actions/setup-node@v3
         with:
-          node-version: 16.18.1
+          node-version: 18.13.0
           registry-url: https://registry.npmjs.org/
           cache: 'yarn'
           cache-depndency-path: '**/yarn.lock'
