+++ .nvmrc
@@ -1,1 +1,1 @@
-16.18.1
+v18.13.0
