+++ package.json
@@ -50,43 +50,43 @@
     "test:ci": "yarn nx affected --target=test:ci"
   },
   "devDependencies": {
-    "@commitlint/cli": "^17.3.0",
-    "@commitlint/config-conventional": "^17.3.0",
-    "@nrwl/devkit": "15.3.0",
-    "@nrwl/jest": "15.3.0",
+    "@commitlint/cli": "^17.4.2",
+    "@commitlint/config-conventional": "^17.4.2",
+    "@nrwl/devkit": "15.5.3",
+    "@nrwl/jest": "15.5.3",
     "@nrwl/nx-cloud": "^15.0.2",
     "@types/eslint": "^8.4.10",
-    "@types/jest": "^29.2.5",
-    "@types/node": "16.18.11",
-    "@typescript-eslint/eslint-plugin": "^5.44.0",
-    "@typescript-eslint/parser": "^5.44.0",
-    "audit-ci": "^6.4.1",
-    "eslint": "^8.28.0",
+    "@types/jest": "^29.2.6",
+    "@types/node": "^18.11.18",
+    "@typescript-eslint/eslint-plugin": "^5.48.2",
+    "@typescript-eslint/parser": "^5.48.2",
+    "audit-ci": "^6.6.0",
+    "eslint": "^8.32.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-airbnb-typescript": "^17.0.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jest": "^27.1.7",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jest": "^27.2.1",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-webpack-plugin": "^3.2.0",
-    "husky": "^8.0.2",
+    "husky": "^8.0.3",
     "is-ci": "^3.0.1",
     "jest": "^29.3.1",
-    "lint-staged": "^13.0.4",
+    "lint-staged": "^13.1.0",
     "nx": "15.3.0",
-    "patch-package": "^6.5.0",
-    "prettier": "^2.8.0",
-    "prettier-plugin-packagejson": "^2.3.0",
+    "patch-package": "^6.5.1",
+    "prettier": "^2.8.3",
+    "prettier-plugin-packagejson": "^2.4.0",
     "ts-node": "^10.9.1",
-    "typescript": "^4.9.3",
+    "typescript": "^4.9.4",
     "webpack": "^5.75.0"
   },
   "packageManager": "yarn@3.3.0",
   "engines": {
-    "node": ">=16.18"
+    "node": ">=18.13.0"
   },
   "volta": {
-    "node": "16.18.1",
+    "node": "18.13.0",
     "yarn": "3.3.0"
   }
 }
