+++ packages/docs/package.json
@@ -44,8 +44,8 @@
     "@ethersproject/providers": "^5.7.2",
     "@mdx-js/react": "1.6.22",
     "@metamask/providers": "^10.2.1",
-    "@mui/icons-material": "^5.10.15",
-    "@mui/material": "^5.10.15",
+    "@mui/icons-material": "^5.11.0",
+    "@mui/material": "^5.11.5",
     "@web3-react/core": "^6.1.9",
     "@web3-react/injected-connector": "^6.0.7",
     "clsx": "^1.2.1",
@@ -58,28 +58,28 @@
   "devDependencies": {
     "@docusaurus/module-type-aliases": "^2.2.0",
     "@tsconfig/docusaurus": "^1.0.6",
-    "@types/react": "18.0.25",
-    "@typescript-eslint/eslint-plugin": "^5.44.0",
-    "@typescript-eslint/parser": "^5.44.0",
-    "eslint": "^8.28.0",
+    "@types/react": "18.0.27",
+    "@typescript-eslint/eslint-plugin": "^5.48.2",
+    "@typescript-eslint/parser": "^5.48.2",
+    "eslint": "^8.32.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^17.0.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jsx-a11y": "^6.6.1",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-react": "^7.31.11",
+    "eslint-plugin-react": "^7.32.1",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "husky": "^8.0.2",
-    "lint-staged": "^13.0.4",
+    "husky": "^8.0.3",
+    "lint-staged": "^13.1.0",
     "precompress": "^9.1.1",
-    "prettier": "^2.8.0",
+    "prettier": "^2.8.3",
     "rimraf": "^3.0.2",
-    "stylelint": "^14.15.0",
+    "stylelint": "^14.16.1",
     "stylelint-config-css-modules": "^4.1.0",
     "stylelint-config-prettier": "^9.0.4",
     "stylelint-config-standard": "^29.0.0",
-    "typescript": "^4.9.3"
+    "typescript": "^4.9.4"
   },
   "nx": {
     "namedInputs": {
