+++ packages/snap/jest.config.ts
@@ -92,7 +92,10 @@ export default {
   // ],
 
   // A map from regular expressions to module names or to arrays of module names that allow to stub out resources with a single module
-  // moduleNameMapper: {},
+  moduleNameMapper: {
+    'multiformats/bases/base58':
+      '<rootDir>/node_modules/multiformats/bases/base58.js',
+  },
 
   // An array of regexp pattern strings, matched against all module paths before considered 'visible' to the module loader
   // modulePathIgnorePatterns: [],
