+++ packages/snap/package.json
@@ -49,7 +49,7 @@
     "test:ci": "yarn test --silent --coverage"
   },
   "dependencies": {
-    "@babel/runtime": "7.20.5",
+    "@babel/runtime": "7.20.13",
     "@blockchain-lab-um/ssi-snap-types": "*",
     "@blockchain-lab-um/veramo-vc-manager": "*",
     "@ceramicnetwork/http-client": "2.11.0",
@@ -90,53 +90,53 @@
     "key-did-provider-ed25519": "^2.0.1",
     "key-did-resolver": "2.3.0",
     "lodash.clonedeep": "^4.5.0",
-    "multiformats": "^11.0.0",
+    "multiformats": "^11.0.1",
     "secp256k1": "4.0.3",
-    "ts-jest": "^29.0.3",
+    "ts-jest": "^29.0.5",
     "uuid": "^9.0.0",
     "web-did-resolver": "2.0.21"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.5",
+    "@babel/core": "^7.20.12",
     "@babel/plugin-proposal-class-properties": "^7.18.6",
-    "@babel/plugin-proposal-decorators": "^7.20.5",
+    "@babel/plugin-proposal-decorators": "^7.20.13",
     "@babel/plugin-proposal-private-methods": "^7.18.6",
     "@babel/plugin-transform-flow-strip-types": "^7.19.0",
     "@babel/preset-typescript": "^7.18.6",
     "@babel/register": "^7.18.9",
-    "@commitlint/cli": "^17.3.0",
-    "@commitlint/config-conventional": "^17.3.0",
+    "@commitlint/cli": "^17.4.2",
+    "@commitlint/config-conventional": "^17.4.2",
     "@metamask/snaps-cli": "^0.27.1",
     "@metamask/snaps-webpack-plugin": "^0.27.1",
-    "@types/jest": "^29.2.5",
+    "@types/jest": "^29.2.6",
     "@types/jsonpath": "^0.2.0",
     "@types/lodash.clonedeep": "^4.5.7",
-    "@types/node": "16.18.11",
-    "@typescript-eslint/eslint-plugin": "^5.44.0",
-    "@typescript-eslint/parser": "^5.44.0",
-    "audit-ci": "^6.3.0",
-    "babel-loader": "^9.1.0",
-    "eslint": "^8.28.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jest": "^27.1.6",
+    "@types/node": "18.11.18",
+    "@typescript-eslint/eslint-plugin": "^5.48.2",
+    "@typescript-eslint/parser": "^5.48.2",
+    "audit-ci": "^6.6.0",
+    "babel-loader": "^9.1.2",
+    "eslint": "^8.32.0",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jest": "^27.2.1",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-webpack-plugin": "^3.2.0",
-    "fork-ts-checker-webpack-plugin": "^7.2.13",
+    "fork-ts-checker-webpack-plugin": "^7.3.0",
     "hook-shell-script-webpack-plugin": "^0.1.4",
-    "husky": "^8.0.2",
+    "husky": "^8.0.3",
     "is-ci": "^3.0.1",
     "jest": "^29.3.1",
-    "lint-staged": "^13.0.4",
+    "lint-staged": "^13.1.0",
     "postinstall-postinstall": "^2.1.0",
     "precompress": "^9.1.1",
-    "prettier": "^2.8.0",
-    "prettier-plugin-packagejson": "^2.3.0",
+    "prettier": "^2.8.3",
+    "prettier-plugin-packagejson": "^2.4.0",
     "rimraf": "^3.0.2",
     "ts-node": "^10.9.1",
-    "typescript": "^4.9.3",
+    "typescript": "^4.9.4",
     "webpack": "^5.75.0",
-    "webpack-cli": "^5.0.0",
+    "webpack-cli": "^5.0.1",
     "webpack-merge": "^5.8.0",
     "webpackbar": "^5.0.2"
   },
