+++ packages/snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/blockchain-lab-um/ssi-snap.git"
   },
   "source": {
-    "shasum": "SFP566pqm5Coy0EehxwozmNgLfBAso5V/IbnsylQRfw=",
+    "shasum": "Vix1oGpYrJa7+EYIgncaxdGiz2SK8IOnLq8rGfHlQ1k=",
     "location": {
       "npm": {
         "filePath": "dist/snap.js",
