+++ packages/types/src/constants.ts
@@ -5,13 +5,13 @@ const isIn = <T>(values: readonly T[], value: any): value is T => {
 };
 
 export const availableVCStores = ['snap', 'ceramic'] as const;
-export type AvailableVCStores = typeof availableVCStores[number];
+export type AvailableVCStores = (typeof availableVCStores)[number];
 
 export const isAvailableVCStores = (x: string) =>
   isIn<AvailableVCStores[number]>(availableVCStores, x);
 
 export const availableMethods = ['did:ethr', 'did:key', 'did:pkh'] as const;
-export type AvailableMethods = typeof availableMethods[number];
+export type AvailableMethods = (typeof availableMethods)[number];
 export const isAvailableMethods = (x: string) =>
   isIn<AvailableMethods[number]>(availableMethods, x);
 
@@ -27,7 +27,7 @@ export const supportedProofFormats = [
   'EthereumEip712Signature2021',
 ] as const;
 
-export type SupportedProofFormats = typeof supportedProofFormats[number];
+export type SupportedProofFormats = (typeof supportedProofFormats)[number];
 
 export const isSupportedProofFormat = (x: string) =>
   isIn(supportedProofFormats, x);
