+++ packages/vcmanager/package.json
@@ -30,20 +30,20 @@
     "uuid": "^9.0.0"
   },
   "devDependencies": {
-    "@types/jest": "^29.2.5",
-    "@types/prettier": "^2.7.1",
+    "@types/jest": "^29.2.6",
+    "@types/prettier": "^2.7.2",
     "@types/rimraf": "^3.0.2",
     "@types/uuid": "^8.3.4",
     "@veramo/cli": "4.2.0",
-    "eslint": "^8.28.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.26.0",
+    "eslint": "^8.32.0",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-prettier": "^4.2.1",
     "jest": "29.3.1",
-    "prettier": "^2.8.0",
+    "prettier": "^2.8.3",
     "rimraf": "^3.0.2",
-    "ts-jest": "29.0.3",
-    "typescript": "^4.9.3"
+    "ts-jest": "29.0.5",
+    "typescript": "^4.9.4"
   },
   "publishConfig": {
     "access": "public"
