+++ packages/vcmanager/src/types/IDataManager.ts
@@ -1,4 +1,5 @@
 import { IPluginMethodMap } from '@veramo/core';
+
 export interface IDataManager extends IPluginMethodMap {
   query(args: IDataManagerQueryArgs): Promise<Array<IDataManagerQueryResult>>;
