+++ packages/website/package.json
@@ -15,28 +15,28 @@
     "@blockchain-lab-um/ssi-snap-types": "*",
     "axios": "^1.1.3",
     "ethers": "^5.7.2",
-    "pinia": "^2.0.21",
+    "pinia": "^2.0.29",
     "primeflex": "^3.3.0",
     "primeicons": "^6.0.1",
-    "primevue": "^3.18.1",
-    "sass": "^1.56.1",
-    "vue": "^3.2.38",
-    "vue-router": "^4.1.5"
+    "primevue": "^3.22.3",
+    "sass": "^1.57.1",
+    "vue": "^3.2.45",
+    "vue-router": "^4.1.6"
   },
   "devDependencies": {
     "@rushstack/eslint-patch": "^1.2.0",
-    "@types/node": "16.18.11",
+    "@types/node": "^18.11.18",
     "@vitejs/plugin-vue": "^3.2.0",
     "@vue/eslint-config-prettier": "^7.0.0",
     "@vue/eslint-config-typescript": "^11.0.2",
     "@vue/tsconfig": "^0.1.3",
-    "eslint": "^8.28.0",
-    "eslint-plugin-vue": "^9.8.0",
+    "eslint": "^8.32.0",
+    "eslint-plugin-vue": "^9.9.0",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.8.0",
-    "typescript": "~4.9.3",
-    "vite": "^3.2.4",
-    "vue-tsc": "^1.0.9"
+    "prettier": "^2.8.3",
+    "typescript": "~4.9.4",
+    "vite": "^3.2.5",
+    "vue-tsc": "^1.0.24"
   },
   "nx": {
     "namedInputs": {
